# Conflicts:
#	core/droidfeature/feature_extraction.py

+++ core/attack/gdkdel1.py
@@ -71,7 +71,7 @@ class GDKDEl1(BaseAttack):
         Parameters
         -----------
         @param model, a victim model
-        @param x: torch.DoubleTensor, feature vectors with shape [batch_size, vocab_dim]
+        @param x: torch.FloatTensor, feature vectors with shape [batch_size, vocab_dim]
         @param label: torch.LongTensor, ground truth labels
         @param m: Integer, maximum number of perturbations
         @param lambda_, float, penalty factor
@@ -110,7 +110,7 @@ class GDKDEl1(BaseAttack):
             self.lambda_ = min_lambda_
         else:
             self.lambda_ = max_lambda_
-        adv_x = x.detach().clone().to(torch.double)
+        adv_x = x.detach().clone()
         while self.lambda_ <= max_lambda_:
             _, done = self.get_loss(model, adv_x, label)
             if torch.all(done):
@@ -152,7 +152,7 @@ class GDKDEl1(BaseAttack):
         # look for important position
         absolute_grad = torch.abs(gradients).reshape(features.shape[0], -1)
         _, position = torch.max(absolute_grad, dim=-1)
-        perturbations = F.one_hot(position, num_classes=absolute_grad.shape[-1]).double()
+        perturbations = F.one_hot(position, num_classes=absolute_grad.shape[-1])
         perturbations = perturbations.reshape(features.shape)
         directions = torch.sign(gradients) * (perturbations > 1e-6)
