+++ examples/max_test.py
@@ -131,7 +131,7 @@ def _main():
                                        name=args.model_name,
                                        **hp_params
                                        )
-    model = model.to(dv).double()
+    model = model.to(dv)
     if args.model == 'md_at_pgd':
         at_wrapper = PGDAdvTraining(model)
         at_wrapper.load()
@@ -156,7 +156,7 @@ def _main():
                                      name=args.model_name,
                                      **hp_params
                                      )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_dnn_plus':
         model = AMalwareDetectionDNNPlus(md_nn_model=None,
@@ -166,7 +166,7 @@ def _main():
                                          name=args.model_name,
                                          **hp_params
                                          )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_pad_ma':
         adv_model = AMalwareDetectionPAD(model)
@@ -174,7 +174,7 @@ def _main():
         model = adv_model.model
     else:
         model.load()
-        model = model.to(dv).double()
+        model = model.to(dv)
     logger.info("Load model parameters from {}.".format(model.model_save_path))
     model.predict(mal_test_dataset_producer)
 
@@ -249,7 +249,7 @@ def _main():
     x_mod_integrated = []
     for x, y in mal_test_dataset_producer:
         x, y = utils.to_tensor(x, y.long(), model.device)
-        adv_x_batch = attack.perturb(model, x.double(), y,
+        adv_x_batch = attack.perturb(model, x, y,
                                      steps_max=args.steps_max,
                                      min_lambda_=1e-5,
                                      max_lambda_=1e5,
