+++ examples/md_at_pgd_test.py
@@ -43,7 +43,7 @@ def _main():
                                 name=model_name,
                                 **vars(args)
                                 )
-    model = model.to(dv).double()
+    model = model.to(dv)
 
     attack = PGD(norm='linf', use_random=False,
                  is_attacker=False,
